+++ Changelog.md
@@ -4,6 +4,7 @@
 
   * Fixed #42 issue.
   * Fixed #44 issue.
+  * #42. If bot nickname already in use, generate new name and try to reconnect.
 
 ## Ybot-0.3.1 --> released